requirements.txt
fastapi
uvicorn
python-dotenv
requests
dnspython
llama-cpp-python